Update convert_matrix function
Currently, `convert_matrix` will raise an error if any column is logical. This is annoying because often design matrices contain indicator variables which may be stored as numeric or logical vectors.

This commit updates `convert_matrix` to not throw an error if there are any logical columns, and to convert logical columns to numeric prior to running the `*model.matrix()` functions.