Added condition to `nn_ind` to capture K
I added a condition to capture old `K` arguments supplied to step_knnimpute(). I encountered the error "$ Operator is invalid for atomic vectors" since object$neighbors did not exist for a recipe I prepped prior to the implementation of https://github.com/tidymodels/recipes/issues/177. It would be a hassle to completely retrain the data. I was able to informally fix the issue with: model$recipe$steps[[2]]$neighbors <- 5, but this proposed change makes it more formal.